#450 arcbiosubsan: ten uncast shifts, and the author had already cast four of them
src/promemul/arcbios.c.  buf[] is unsigned char; buf[3] << 24 promotes to int, and with
buf[3] >= 0x80 the result is unrepresentable -- C99 6.5.7p4.  FOURTEEN sites, FOUR already
cast correctly -- line 580 sits between two uncast siblings at 570 and 590, so the author
knew the idiom and applied it to 4 of 14.  A class, not a typo.

At the 64-bit-assembly sites the wrapped-negative int converts to uint64_t before the
high-byte terms are added, so the assembled pointer is V - 2^32: LOW WORD CORRECT, HIGH
WORD ONE TOO SMALL.  Witnessed value: FIRST_ARC_COMPONENT = 0xffffffffbfca8000, whose
swapped buf[3] is 0xBF = 191 -- the exact number in the UBSan message.

MEASURED pre-fix on a freshly rebuilt instrumented tree: 2 shift lines on each reaching
subtype (:570/:590 on ip12/ip32/ip35, :737/:784 on ip28/ip30), ASan 0 -- the rebuild also
independently re-confirmed #446.  Post-fix: shift=0 asan=0 on all ten subtypes.

SCOPE WAS CONTESTED IN PASS 1 AND ADJUDICATED ALL TEN.  One seat argued the four witnessed
sites under the round rule; two argued the class.  The decisive form: THE CAST IS PROVABLY
VALUE-NEUTRAL WHEREVER THE UB IS NOT FIRING, so there is nothing per-site for a witness to
adjudicate -- and the coverage objection (six fixes no sweep can exercise) is answered by
the census row, PROVEN by kill-table row M2: reverting an unwitnessed cast leaves the sweep
green and reddens ONLY the census.

Width-matched casts per the house idiom, whose true split a seat corrected: the four
pre-existing casts are 3-to-1 ((uint64_t) at :1349/:1400/:1449, (uint32_t) only at :580),
not 2-and-2 as the brief claimed.

*** A VOID MEASUREMENT CAUGHT MID-ROUND: build_asan.sh BUILDS FROM GIT HEAD. ***  Its
first "post-fix" sweep reported the PRE-fix counts with rc=0, because `git archive HEAD`
cannot see an uncommitted edit.  Rule 4 caught it -- the UBSan lines still named the OLD
sites -- and the real verification used a private tree with the working-tree file installed
and VERIFIED BY CMP before building.  rc=0 from that script means "HEAD built", never
"your edit built".

TWO GATE-9 ROWS, because the existing three-way CANNOT pin this: sanhit() already counted
these `runtime error:` lines, but the gate's assertion is DIRECTIONAL and a revert lands on
dirty/dirty -- "pre-existing, reported but not failed".  Invisible.

*** PASS 2: FOUR SEATS KILLED THE CENSUS V1 WITH ONE TOKEN EACH. ***  The first census was
two greps.  deepseek: `& 0x7fffffff` after a cast (spelling kept, sanitizers silent, bit 31
gone).  codex: `~(uint32_t)` one character earlier, and a comma-expression smuggling a
signed shift behind a cast match.  glm: ANY signed cast ending in `)` evaded both regexes,
and an ADDED dead `(int)buf[3]<<24` statement reintroduced the exact UB while both counts
held.  grok: a suffix on a still-cast shift, same class.  A PATTERN PINS WHAT ITS AUTHOR
THOUGHT OF.

V2 is regress/arcbios_shift_probe.py -- the #444 S4 precedent: every buf[]-assembly
statement, comment/string-stripped and whitespace-normalised, exact-matched as a MULTISET
against a frozen table of 14 canonical statements (GENERATED from the reviewed file, not
typed: the hand-assembled-encoding trap has five recorded incidents).  MEASURED: all four
seat escapes plus the earlier three mutants die to C1; an added statement is caught as a
multiset EXTRA; a comment quoting the uncast idiom no longer false-positives.  The accepted
S4 trade: a legitimate edit to any of the fourteen must update the table in the same
commit.  There is deliberately NO separate count row -- six mutants all died to C1, and a
row that can never fire alone is decoration.

THE BEHAVIOURAL ROW GAINED A LIVENESS TERM after codex named the hole: zero matching lines
from a run that never CONSTRUCTED is an absent measurement, so each subtype must show its
own `model: SGI-` banner or the row reddens.  *** THE FIRST TOKEN CHOSEN WAS ITSELF
MEASURED WRONG BEFORE SHIPPING *** -- it returned 0 on a live construction because the
probe run lacked gate 9's detect_leaks=0 environment; verified present on all five under
the gate's exact env before landing.

THE ROUND'S OWN PASS-1 ANALYSIS WAS CORRECTED TWICE BY THE PANEL: the claimed "mis-linked
component tree" was a COUNTERFACTUAL about :751 (the witnessed :737 feeds only == 0 tests,
which corruption cannot alias -- it forces the low word >= 2^31); and addchild64's parent
parameter is uint64_t, not a promoted uint32_t.  A seat completed the trace: BOTH mis-link
directions exist (false negative at true high word 0, false positive at 1), data-dependent.
The Get* trio is value-benign because the 4-byte branch and the (int64_t)(int32_t)
truncation are keyed on the SAME arc_64bit condition -- read and confirmed.

RESIDUAL FILED, NOT FIXED: arcbioswalk64.  The 64-bit walker reads peeraddr+0x34 for
"IdentifierLength" -- with the 0x18-byte pointer header that offset is INSIDE AffinityMask
(+0x30..0x37); IdentifierLength is at +0x40.  The walker masks the mis-read and skips
~1 MiB per node, so the walk leaves the list after the first node.  Grok completed the
irony: that mis-read is WHY :784 is witnessed at all -- AffinityMask 0xffffffff supplies
the buf[3] >= 0x80.  One-way coupling recorded: fixing the walker may make :751 live.

Also: the deepseek Ollama alias was refreshed preview -> release by owner directive
(`ollama pull deepseek-v4-pro:cloud`); the wire tag is unchanged and the seat
health-verified the same hour.

Gates: census + kill tables measured pre-commit; gate 9's behavioural row requires the
committed tree by construction and runs next.  Twin tree and both NO-VPATH trees carry
arcbios.c byte-identically; build/ rebuilt, rc=0, zero warnings.

+++ src/promemul/arcbios.c
@@ -567,7 +567,7 @@ static uint64_t arcbios_addchild(struct cpu *cpu,
 			tmp2 = buf[0]; buf[0] = buf[3]; buf[3] = tmp2;
 			tmp2 = buf[1]; buf[1] = buf[2]; buf[2] = tmp2;
 		}
-		epeer   = buf[0] + (buf[1]<<8) + (buf[2]<<16) + (buf[3]<<24);
+		epeer   = buf[0] + (buf[1]<<8) + (buf[2]<<16) + ((uint32_t)buf[3]<<24);
 
 		cpu->memory_rw(cpu, cpu->mem, peeraddr + 1 *
 		    machine->md.arc->wordlen,
@@ -587,7 +587,7 @@ static uint64_t arcbios_addchild(struct cpu *cpu,
 			tmp2 = buf[0]; buf[0] = buf[3]; buf[3] = tmp2;
 			tmp2 = buf[1]; buf[1] = buf[2]; buf[2] = tmp2;
 		}
-		eparent = buf[0] + (buf[1]<<8) + (buf[2]<<16) + (buf[3]<<24);
+		eparent = buf[0] + (buf[1]<<8) + (buf[2]<<16) + ((uint32_t)buf[3]<<24);
 
 		/*  debug("  epeer=%x echild=%x eparent=%x\n",
 		    (int)epeer,(int)echild,(int)eparent);  */
@@ -616,7 +616,7 @@ static uint64_t arcbios_addchild(struct cpu *cpu,
 			tmp2 = buf[1]; buf[1] = buf[2]; buf[2] = tmp2;
 		}
 		
-		tmp = buf[0] + (buf[1]<<8) + (buf[2]<<16) + (buf[3]<<24);
+		tmp = buf[0] + (buf[1]<<8) + (buf[2]<<16) + ((uint32_t)buf[3]<<24);
 		peeraddr += 0x30;
 		peeraddr += tmp + 1;
 		peeraddr = ((peeraddr - 1) | 3) + 1;
@@ -720,7 +720,7 @@ static uint64_t arcbios_addchild64(struct cpu *cpu,
 			tmp2 = buf[2]; buf[2] = buf[5]; buf[5] = tmp2;
 			tmp2 = buf[3]; buf[3] = buf[4]; buf[4] = tmp2;
 		}
-		epeer   = buf[0] + (buf[1]<<8) + (buf[2]<<16) + (buf[3]<<24)
+		epeer   = buf[0] + (buf[1]<<8) + (buf[2]<<16) + ((uint64_t)buf[3]<<24)
 		    + ((uint64_t)buf[4] << 32) + ((uint64_t)buf[5] << 40)
 		    + ((uint64_t)buf[6] << 48) + ((uint64_t)buf[7] << 56);
 
@@ -734,7 +734,7 @@ static uint64_t arcbios_addchild64(struct cpu *cpu,
 			tmp2 = buf[2]; buf[2] = buf[5]; buf[5] = tmp2;
 			tmp2 = buf[3]; buf[3] = buf[4]; buf[4] = tmp2;
 		}
-		echild  = buf[0] + (buf[1]<<8) + (buf[2]<<16) + (buf[3]<<24)
+		echild  = buf[0] + (buf[1]<<8) + (buf[2]<<16) + ((uint64_t)buf[3]<<24)
 		    + ((uint64_t)buf[4] << 32) + ((uint64_t)buf[5] << 40)
 		    + ((uint64_t)buf[6] << 48) + ((uint64_t)buf[7] << 56);
 
@@ -748,7 +748,7 @@ static uint64_t arcbios_addchild64(struct cpu *cpu,
 			tmp2 = buf[2]; buf[2] = buf[5]; buf[5] = tmp2;
 			tmp2 = buf[3]; buf[3] = buf[4]; buf[4] = tmp2;
 		}
-		eparent = buf[0] + (buf[1]<<8) + (buf[2]<<16) + (buf[3]<<24)
+		eparent = buf[0] + (buf[1]<<8) + (buf[2]<<16) + ((uint64_t)buf[3]<<24)
 		    + ((uint64_t)buf[4] << 32) + ((uint64_t)buf[5] << 40)
 		    + ((uint64_t)buf[6] << 48) + ((uint64_t)buf[7] << 56);
 
@@ -781,7 +781,7 @@ static uint64_t arcbios_addchild64(struct cpu *cpu,
 			tmp2 = buf[0]; buf[0] = buf[3]; buf[3] = tmp2;
 			tmp2 = buf[1]; buf[1] = buf[2]; buf[2] = tmp2;
 		}
-		tmp = buf[0] + (buf[1]<<8) + (buf[2]<<16) + (buf[3]<<24);
+		tmp = buf[0] + (buf[1]<<8) + (buf[2]<<16) + ((uint64_t)buf[3]<<24);
 
 		tmp &= 0xfffff;
 
@@ -1359,7 +1359,7 @@ int arcbios_emul(struct cpu *cpu)
 					buf[2] = tmp;
 				}
 				peer = buf[0] + (buf[1]<<8) + (buf[2]<<16)
-				    + (buf[3]<<24);
+				    + ((uint64_t)buf[3]<<24);
 			}
 
 			cpu->cd.mips.gpr[MIPS_GPR_V0] = peer?
@@ -1410,7 +1410,7 @@ int arcbios_emul(struct cpu *cpu)
 					buf[2] = tmp;
 				}
 				child = buf[0] + (buf[1]<<8) + (buf[2]<<16) +
-				    (buf[3]<<24);
+				    ((uint64_t)buf[3]<<24);
 			}
 
 			cpu->cd.mips.gpr[MIPS_GPR_V0] = child?
@@ -1459,7 +1459,7 @@ int arcbios_emul(struct cpu *cpu)
 					buf[2] = tmp;
 				}
 				parent = buf[0] + (buf[1]<<8) +
-				    (buf[2]<<16) + (buf[3]<<24);
+				    (buf[2]<<16) + ((uint64_t)buf[3]<<24);
 			}
 
 			cpu->cd.mips.gpr[MIPS_GPR_V0] = parent?
